# docs/openapi.yaml
openapi: 3.1.0
info:
  title: POS System API
  version: 0.1.0
  description: POS health checks and session authentication contract.
servers:
  - url: http://localhost:8080
paths:
  /v1/auth/login:
    post:
      operationId: login
      summary: Create a bearer session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 1, maxLength: 72 }
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, expires_at]
                properties:
                  access_token: { type: string }
                  token_type: { type: string, const: Bearer }
                  expires_at: { type: string, format: date-time }
        '400': { description: Invalid request }
        '401': { description: Invalid credentials }
  /v1/me:
    get:
      operationId: currentUser
      summary: Return the authenticated user and permissions
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user }
        '401': { description: Missing, invalid, or expired bearer token }
  /v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current bearer session
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Session revoked }
        '401': { description: Missing or invalid bearer token }
  /v1/orders/{orderID}/payments:
    post:
      operationId: chargeOrder
      summary: Charge an order using the configured payment provider
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderID
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        '200': { description: Payment attempt result }
        '400': { description: Invalid order identifier or payment response }
        '401': { description: Missing or invalid bearer token }
        '403': { description: User lacks payment.create permission }
        '409': { description: Order is not payable }
        '502': { description: Payment provider unavailable }
  /v1/orders/checkout:
    post:
      operationId: checkout
      summary: Create an order and atomically decrement stock
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [items]
              properties:
                customer_id: { type: integer, format: int64, minimum: 1 }
                items:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    type: object
                    additionalProperties: false
                    required: [product_id, quantity]
                    properties:
                      product_id: { type: integer, format: int64, minimum: 1 }
                      quantity: { type: integer, format: int64, minimum: 1 }
      responses:
        '201': { description: Pending order created with server-derived prices }
        '400': { description: Invalid checkout request }
        '401': { description: Missing or invalid bearer token }
        '403': { description: User lacks order.create permission }
        '404': { description: Product is unavailable }
        '409': { description: Insufficient stock }
  /v1/categories:
    get:
      operationId: listCategories
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Active categories }
        '401': { description: Missing or invalid bearer token }
        '403': { description: User lacks product.read permission }
    post:
      operationId: createCategory
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string, maxLength: 1000 }
      responses:
        '201': { description: Category created }
        '400': { description: Invalid category }
        '401': { description: Missing or invalid bearer token }
        '403': { description: User lacks product.create permission }
        '409': { description: Category name already exists }
  /v1/products:
    get:
      operationId: listProducts
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string, maxLength: 200 } }
        - { name: category_id, in: query, schema: { type: integer, format: int64, minimum: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200': { description: Active products }
        '400': { description: Invalid list filters }
        '401': { description: Missing or invalid bearer token }
        '403': { description: User lacks product.read permission }
    post:
      operationId: createProduct
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [sku, name, price_minor]
              properties:
                category_id: { type: integer, format: int64, minimum: 1 }
                sku: { type: string, minLength: 1, maxLength: 64 }
                name: { type: string, minLength: 1, maxLength: 200 }
                description: { type: string, maxLength: 2000 }
                price_minor: { type: integer, format: int64, minimum: 0 }
                currency: { type: string, pattern: '^[A-Z]{3}$', default: IDR }
      responses:
        '201': { description: Product created }
        '400': { description: Invalid product or category }
        '401': { description: Missing or invalid bearer token }
        '403': { description: User lacks product.create permission }
        '409': { description: SKU already exists }
  /readyz:
    get:
      operationId: readinessCheck
      summary: Check whether the service can access its database
      responses:
        '200':
          description: Database is reachable
          content:
            text/plain:
              schema:
                type: string
              example: ready
        '503':
          description: Database is unavailable
  /healthz:
    get:
      operationId: healthCheck
      summary: Check process health
      responses:
        '200':
          description: The HTTP process is running
          content:
            text/plain:
              schema:
                type: string
              example: ok
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer